+++ app/src/main/java/app/passwordstore/data/repo/PasswordRepository.kt
@@ -33,8 +33,6 @@ object PasswordRepository {
   val isInitialized: Boolean
     get() = repository != null
 
-  var gpgidChecked: Boolean = false
-
   fun isGitRepo(): Boolean {
     return repository?.objectDatabase?.exists() ?: false
   }