+++ app/src/main/java/app/passwordstore/ui/crypto/DecryptActivity.kt
@@ -17,7 +17,6 @@ import app.passwordstore.crypto.errors.IncorrectPassphraseException
 import app.passwordstore.crypto.errors.NoDecryptionKeyAvailableException
 import app.passwordstore.data.passfile.PasswordEntry
 import app.passwordstore.data.password.FieldItem
-import app.passwordstore.data.repo.PasswordRepository
 import app.passwordstore.databinding.DecryptLayoutBinding
 import app.passwordstore.ui.adapters.FieldItemAdapter
 import app.passwordstore.util.extensions.enableEdgeToEdgeView
@@ -63,17 +62,7 @@ class DecryptActivity : BasePGPActivity() {
       }
     }
     requireKeysExist {
-      val gpgIdentifiers =
-        getPGPIdentifiers(relativeParentPath)?.filter { repository.hasKey(it) }
-          ?: return@requireKeysExist
-      if (gpgIdentifiers.isEmpty()) {
-        snackbar(message = resources.getString(R.string.password_decryption_no_decryption_key))
-        return@requireKeysExist
-      } else if (gpgIdentifiers.filter { repository.hasSecretKey(it) }.isEmpty()) {
-        snackbar(message = resources.getString(R.string.password_decryption_no_decryption_key))
-        return@requireKeysExist
-      }
-      getPersistentAndDecrypt(gpgIdentifiers)
+      requireDecryptionKeysExist(relativeParentPath) { ids -> getPersistentAndDecrypt(ids) }
     }
   }
 
@@ -111,7 +100,6 @@ class DecryptActivity : BasePGPActivity() {
       } else if (
         results.filter { it.second.getError() is NoDecryptionKeyAvailableException }.any()
       ) {
-        PasswordRepository.gpgidChecked = false
         snackbar(message = resources.getString(R.string.password_decryption_no_decryption_key))
       } else {
         snackbar(message = resources.getString(R.string.password_decryption_unknown_error))